test 0.5

+++ CarDealerShip/DomainLayer/Kunde.cs
@@ -4,7 +4,7 @@
     {
         
 
-        public Kunde(string name, string adresse, string email, int telefon) : base(adresse, email, name, telefon)
+        public Kunde(string name, string adresse, string email, long telefon) : base(adresse, email, name, telefon)
         {
             
         }
